refactor: move classes in their own package
