+++ store/src/main/java/com/zextras/mailbox/api/rest/resource/PingResource.java
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-package com.zextras.mailbox.api.rest;
+package com.zextras.mailbox.api.rest.resource;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
