+++ store/src/test/java/com/zextras/mailbox/api/resource/AccountResourceIT.java
@@ -1,4 +1,4 @@
-package com.zextras.mailbox.api;
+package com.zextras.mailbox.api.resource;
 
 import static org.junit.jupiter.api.Assertions.*;
