+++ store/src/test/java/com/zextras/mailbox/api/resource/MailboxResourceIT.java
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-package com.zextras.mailbox.api;
+package com.zextras.mailbox.api.resource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
